Move macros to own file
